refactor: (profile)
Fetching user profile moved to summoner component.  basicProfile converted to presentational component.
